# specs/001-matrades-product-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Matrades Application API
  version: 1.1.0
  description: >-
    Contract for authenticated configuration, research, strategy, risk, approval,
    reconciliation, monitoring, journal, and performance workflows. Broker execution
    writes are intentionally absent from V1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /accounts/{account_id}/research-matrix:
    get:
      tags: [Research]
      operationId: getAccountResearchMatrix
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Active versioned asset-class and instrument-type research matrix
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResearchMatrix' }
        default: { $ref: '#/components/responses/Error' }
    put:
      tags: [Research]
      operationId: replaceAccountResearchMatrix
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResearchMatrixInput' }
      responses:
        '200':
          description: New immutable research-matrix version activated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResearchMatrix' }
        '409': { $ref: '#/components/responses/Conflict' }
        default: { $ref: '#/components/responses/Error' }
  /accounts/{account_id}/provider-bindings:
    get:
      tags: [Configuration]
      operationId: listAccountProviderBindings
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Effective capability bindings for the account's research lanes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProviderBinding' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Configuration]
      operationId: createAccountProviderBinding
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderBindingInput' }
      responses:
        '201':
          description: Immutable provider-binding version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderBinding' }
        default: { $ref: '#/components/responses/Error' }
  /instruments:
    get:
      tags: [Research]
      operationId: listInstruments
      parameters:
        - name: asset_class
          in: query
          schema: { $ref: '#/components/schemas/AssetClass' }
        - name: instrument_type
          in: query
          schema: { $ref: '#/components/schemas/InstrumentType' }
        - name: connection_id
          in: query
          schema: { type: string, format: uuid }
        - name: capability
          in: query
          schema: { $ref: '#/components/schemas/MarketDataCapability' }
      responses:
        '200':
          description: Canonical instruments and exact provider or broker listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InstrumentDescriptor' }
        default: { $ref: '#/components/responses/Error' }
...
  /research-runs:
    post:
      tags: [Research]
      operationId: createResearchRun
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id]
              properties:
                account_id: { type: string, format: uuid }
                lanes:
                  type: array
                  description: Omit to run every enabled lane in the account's active matrix
                  minItems: 1
                  maxItems: 12
                  uniqueItems: true
                  items: { $ref: '#/components/schemas/ResearchLaneKey' }
      responses:
        '202':
          description: Research run queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResearchRun' }
        default: { $ref: '#/components/responses/Error' }
  /research-runs/{run_id}:
    get:
      tags: [Research]
      operationId: getResearchRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Ranked candidates, evidence, health, and HIL-1 state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResearchRun' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: matrades_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Expected aggregate/configuration version
      schema: { type: string }
    AccountId:
      name: account_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Structured safe error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Idempotency, state-transition, or optimistic-concurrency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        correlation_id: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    AssetClass:
      type: string
      enum: [FOREX, METALS, CRYPTOCURRENCY, STOCKS]
    InstrumentType:
      type: string
      enum: [SPOT, CFD, FUTURES]
    MarketDataCapability:
      type: string
      enum: [INSTRUMENT_DIRECTORY, DISCOVERY, QUOTE, TRADES, CANDLES, ORDER_BOOK,
        CONTRACT_DETAILS, FUTURES_CHAIN, OPEN_INTEREST, FUNDING, CORPORATE_ACTIONS,
        BROKER_TRADABILITY, ACCOUNT_BALANCES]
    ProviderAuthorityPurpose:
      type: string
      enum: [DISCOVERY, REFERENCE, EXECUTABLE_QUOTE, HISTORY, CONTRACT_TERMS,
        BROKER_RECONCILIATION]
    ResearchLaneKey:
      type: object
      required: [asset_class, instrument_type]
      properties:
        asset_class: { $ref: '#/components/schemas/AssetClass' }
        instrument_type: { $ref: '#/components/schemas/InstrumentType' }
    ResearchMatrixInput:
      type: object
      required: [lanes, schedule]
      properties:
        lanes:
          type: array
          minItems: 1
          maxItems: 12
          uniqueItems: true
          items:
            allOf:
              - $ref: '#/components/schemas/ResearchLaneKey'
              - type: object
                required: [enabled]
                properties:
                  enabled: { type: boolean }
        schedule: { type: object, additionalProperties: true }
    ResearchMatrix:
      allOf:
        - $ref: '#/components/schemas/ResearchMatrixInput'
        - type: object
          required: [id, account_id, version, effective_from]
          properties:
            id: { type: string, format: uuid }
            account_id: { type: string, format: uuid }
            version: { type: integer, minimum: 1 }
            effective_from: { type: string, format: date-time }
    ProviderBindingInput:
      type: object
      required: [lane, capability, authority_purpose, connection_id, priority,
        freshness_policy]
      properties:
        lane: { $ref: '#/components/schemas/ResearchLaneKey' }
        capability: { $ref: '#/components/schemas/MarketDataCapability' }
        authority_purpose: { $ref: '#/components/schemas/ProviderAuthorityPurpose' }
        connection_id: { type: string, format: uuid }
        priority: { type: integer, minimum: 1 }
        provider_venue: { type: string }
        freshness_policy: { type: object, additionalProperties: true }
        effective_from: { type: string, format: date-time }
    ProviderBinding:
      allOf:
        - $ref: '#/components/schemas/ProviderBindingInput'
        - type: object
          required: [id, account_id, version, verification_status]
          properties:
            id: { type: string, format: uuid }
            account_id: { type: string, format: uuid }
            version: { type: integer, minimum: 1 }
            verification_status: { enum: [UNVERIFIED, VERIFIED, STALE, INVALID] }
    InstrumentSpecification:
      type: object
      required: [id, version, effective_from, price_currency, quantity_unit, tick_size,
        quantity_minimum, quantity_step, provenance, freshness]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        effective_from: { type: string, format: date-time }
        effective_to: { type: [string, 'null'], format: date-time }
        price_currency: { type: string }
        settlement_currency: { type: string }
        quantity_unit: { enum: [UNITS, SHARES, LOTS, CONTRACTS] }
        contract_multiplier: { type: string }
        tick_size: { type: string }
        tick_value: { type: string }
        quantity_minimum: { type: string }
        quantity_maximum: { type: string }
        quantity_step: { type: string }
        margin_terms: { type: object, additionalProperties: true }
        financing_terms: { type: object, additionalProperties: true }
        expiry: { type: [string, 'null'], format: date-time }
        first_notice: { type: [string, 'null'], format: date-time }
        last_trade: { type: [string, 'null'], format: date-time }
        provenance: { type: object, additionalProperties: true }
        freshness: { enum: [VALID, SUSPECT, STALE, INVALID] }
    InstrumentDescriptor:
      type: object
      required: [instrument_id, venue_instrument_id, venue, symbol, asset_class,
        instrument_type, specification]
      properties:
        instrument_id: { type: string, format: uuid }
        venue_instrument_id: { type: string, format: uuid }
        venue: { type: string }
        symbol: { type: string }
        asset_class: { $ref: '#/components/schemas/AssetClass' }
        instrument_type: { $ref: '#/components/schemas/InstrumentType' }
        futures_contract_id: { type: [string, 'null'], format: uuid }
        continuous_series: { type: boolean, default: false }
        executable: { type: boolean, default: true }
        specification: { $ref: '#/components/schemas/InstrumentSpecification' }
    ResearchCandidate:
      type: object
      required: [candidate_id, lane, instrument, rank, evidence_refs]
      properties:
        candidate_id: { type: string, format: uuid }
        lane: { $ref: '#/components/schemas/ResearchLaneKey' }
        instrument: { $ref: '#/components/schemas/InstrumentDescriptor' }
        rank: { type: integer, minimum: 1 }
        evidence_refs: { type: array, items: { type: string } }
    ResearchLaneResult:
      type: object
      required: [lane, status, completed_at]
      properties:
        lane: { $ref: '#/components/schemas/ResearchLaneKey' }
        status: { enum: [READY, NO_TRADE, NOT_CONFIGURED, UNAVAILABLE, STALE, BLOCKED] }
        candidate: { $ref: '#/components/schemas/ResearchCandidate' }
        exclusions: { type: array, items: { type: object, additionalProperties: true } }
        binding_id: { type: string, format: uuid }
        source_cut_refs: { type: array, items: { type: string } }
        reason_code: { type: string }
        completed_at: { type: string, format: date-time }
    ResearchRun:
      type: object
      required: [id, account_id, matrix_version, status, lane_results, version]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        matrix_version: { type: integer, minimum: 1 }
        status: { enum: [QUEUED, RESEARCHING, MARKETS_PENDING_APPROVAL, MARKETS_APPROVED,
          DEGRADED, NO_TRADE, FAILED, EXPIRED] }
        lane_results:
          type: array
          minItems: 1
          maxItems: 12
          items: { $ref: '#/components/schemas/ResearchLaneResult' }
        version: { type: integer }
...